+++ parquet_file/src/metadata.rs
@@ -86,13 +86,14 @@
 //! [Apache Parquet]: https://parquet.apache.org/
 //! [Apache Thrift]: https://thrift.apache.org/
 //! [Thrift Compact Protocol]: https://github.com/apache/thrift/blob/master/doc/specs/thrift-compact-protocol.md
-use std::sync::Arc;
+use std::{convert::TryInto, sync::Arc};
 
 use data_types::partition_metadata::{
     ColumnSummary, InfluxDbType, StatValues, Statistics, TableSummary,
 };
 use internal_types::schema::{InfluxColumnType, InfluxFieldType, Schema};
 use parquet::{
+    arrow::parquet_to_arrow_schema,
     file::{
         metadata::{
             FileMetaData as ParquetFileMetaData, ParquetMetaData,
@@ -177,6 +178,16 @@ pub enum Error {
         column: String,
         source: parquet::errors::ParquetError,
     },
+
+    #[snafu(display("Cannot read arrow schema from parquet: {}", source))]
+    ArrowFromParquetFailure {
+        source: parquet::errors::ParquetError,
+    },
+
+    #[snafu(display("Cannot read IOx schema from arrow: {}", source))]
+    IoxFromArrowFailure {
+        source: internal_types::schema::Error,
+    },
 }
 pub type Result<T, E = Error> = std::result::Result<T, E>;
 
@@ -185,6 +196,24 @@ pub fn read_parquet_metadata_from_file(data: Vec<u8>) -> Result<ParquetMetaData>
     let cursor = SliceableCursor::new(data);
     let reader = SerializedFileReader::new(cursor).context(ParquetMetaDataRead {})?;
     Ok(reader.metadata().clone())
+}
+
+/// Read IOx schema from parquet metadata.
+pub fn read_schema_from_parquet_metadata(parquet_md: &ParquetMetaData) -> Result<Schema> {
+    let file_metadata = parquet_md.file_metadata();
+
+    let arrow_schema = parquet_to_arrow_schema(
+        file_metadata.schema_descr(),
+        file_metadata.key_value_metadata(),
+    )
+    .context(ArrowFromParquetFailure {})?;
+
+    let arrow_schema_ref = Arc::new(arrow_schema);
+
+    let schema: Schema = arrow_schema_ref
+        .try_into()
+        .context(IoxFromArrowFailure {})?;
+    Ok(schema)
 }
 
 /// Read IOx statistics (including timestamp range) from parquet metadata.
@@ -444,9 +473,8 @@ mod tests {
 
     use internal_types::{schema::TIME_COLUMN_NAME, selection::Selection};
 
-    use crate::{
-        storage::read_schema_from_parquet_metadata,
-        utils::{load_parquet_from_store, make_chunk, make_chunk_no_row_group, make_object_store},
+    use crate::utils::{
+        load_parquet_from_store, make_chunk, make_chunk_no_row_group, make_object_store,
     };
 
     #[tokio::test]
